refactor(profile): split profile page into hook + section components
Break the ~610-line pages/[profile].tsx god-component into a lean
orchestrator plus focused, single-responsibility pieces. Behaviour is
unchanged; both the classic and minimal layouts render identically.

Extracted:
- hooks/useProfileAnalytics.ts — view recording + recordClick helper.
- components/profile/ClassicProfile.tsx — composes the classic layout.
- components/profile/ProfileHero.tsx — avatar, name, bio, socials, CV.
- components/profile/ProfileSkills.tsx — GitHub graph, tech stack, about.
- components/profile/ProjectsShowcase.tsx — project grid + CTA.
- components/profile/ShareButton.tsx — floating share button.

pages/[profile].tsx now only handles data loading (getServerSideProps),
theme resolution, <Head>, the layout branch, and the share modal
(~610 -> ~180 lines).
